refactor: renamed app from sample to wildex
