+++ app/src/main/java/com/android/wildex/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.wildex
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -12,19 +12,19 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.tooling.preview.Preview
-import com.android.sample.resources.C
-import com.android.sample.ui.theme.SampleAppTheme
+import com.android.wildex.resources.C
+import com.android.wildex.ui.theme.SampleAppTheme
 
-class SecondActivity : ComponentActivity() {
+class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     setContent {
       SampleAppTheme {
         // A surface container using the 'background' color from the theme
         Surface(
-            modifier = Modifier.fillMaxSize().semantics { testTag = C.Tag.second_screen_container },
+            modifier = Modifier.fillMaxSize().semantics { testTag = C.Tag.main_screen_container },
             color = MaterialTheme.colorScheme.background) {
-              GreetingRobo("Robolectric")
+              Greeting("Android")
             }
       }
     }
@@ -32,12 +32,12 @@ class SecondActivity : ComponentActivity() {
 }
 
 @Composable
-fun GreetingRobo(name: String, modifier: Modifier = Modifier) {
-  Text(text = "Hello $name!", modifier = modifier.semantics { testTag = C.Tag.greeting_robo })
+fun Greeting(name: String, modifier: Modifier = Modifier) {
+  Text(text = "Hello $name!", modifier = modifier.semantics { testTag = C.Tag.greeting })
 }
 
 @Preview(showBackground = true)
 @Composable
-fun GreetingPreview2() {
-  SampleAppTheme { GreetingRobo("Robolectric") }
+fun GreetingPreview() {
+  SampleAppTheme { Greeting("Android") }
 }
