+++ app/src/main/java/com/android/wildex/resources/C.kt
@@ -1,4 +1,4 @@
-package com.android.sample.resources
+package com.android.wildex.resources
 
 // Like R, but C
 object C {
