+++ app/src/test/java/com/android/wildex/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.wildex
 
 import org.junit.Assert.*
 import org.junit.Test
